Creation of the package "network"
Nothing work for now.
The idea is to have a GenericDao with generic methods always coming back
for every entity (eg getById and stuff like that)
Then this generic dao is extended according to the need of the entity
(eg for the users, we also want a method to get the friends)

A special object "Scalingo" have been added, the ultime goal is to
represent an abstraction of the remote server. It contains helper
methods for generating requests. The Daos will keep of reference to it
to be able to send the network requests

The ScalingoModel is an interface implemented by the model objects
travelling on the network. For example, it specifies a special method
"jsonify" to create manually the json representation of an object.